fix removing bug and complete refactoring
