+++ AlexanderMandrov/Lesson15/index.html
@@ -65,6 +65,7 @@
     <script src="./js/canvas.js" type="module"></script>
     <script src="./js/services.js" type="module"></script>
     <script src="./js/Basket.js" type="module"></script>
+    <script src="./js/init.js" type="module"></script>
     <script src="./js/shop.js" type="module"></script>
 </body>
 </html>
