+++ AlexanderMandrov/Lesson15/js/init.js
@@ -1,0 +1,5 @@
+import ItemsList from './ItemsList.js'
+import initCanvasDrawing from  './canvas.js';
+
+initCanvasDrawing();
+(new ItemsList()).fetchItems();
